Refactor : local-storage 모듈화
<description>
많이 쓰일 것 같아서 따로 분리했다.
