+++ src/index.js
@@ -4,6 +4,10 @@ import {addStationScreen} from './View/add-screen.js';
 import {removeStationScreen} from './View/remove-screen.js';
 import {$stationAddInput} from './View/input.js';
 import {isInputValid} from './Controller/valid.js';
+import {
+  setLocalStorage,
+  removeLocalStorage,
+} from './Controller/local-storage.js';
 
 export function onChangeScreen(e) {
   hideScreen();
@@ -12,7 +16,7 @@ export function onChangeScreen(e) {
 
 export function onAddStation() {
   if (isInputValid($stationAddInput.value)) {
-    setLocalStorage('station');
+    setLocalStorage('station', $stationAddInput.value);
     addStationScreen($stationAddInput.value);
   }
   $stationAddInput.value = '';
@@ -22,30 +26,3 @@ export function onRemoveStation(e) {
   removeLocalStorage('station', e.target.dataset.station);
   removeStationScreen(e.target);
 }
-
-export const setLocalStorage = (key) => {
-  const localStorageValue = getLocalStorage(key);
-  if (localStorageValue === null) {
-    return localStorage.setItem(key, JSON.stringify([$stationAddInput.value]));
-  }
-
-  return localStorage.setItem(
-    key,
-    JSON.stringify([...localStorageValue, $stationAddInput.value]),
-  );
-};
-
-export const removeLocalStorage = (key, value) => {
-  const localStorageValue = getLocalStorage(key);
-  const filteredStorage = localStorageValue.filter(
-    (station) => station !== value,
-  );
-
-  return localStorage.setItem(key, JSON.stringify(filteredStorage));
-};
-
-export const getLocalStorage = (key) => {
-  const localStorageValue = JSON.parse(localStorage.getItem(key));
-
-  return localStorageValue;
-};
